+++ java/javasrc/src/main/java/com/longport/quote/OptionVolumeDaily.java
@@ -1,5 +1,6 @@
 package com.longport.quote;
 
 public class OptionVolumeDaily {
+    public String symbol;
     public OptionVolumeDailyStat[] stats;
 }