+++ java/javasrc/src/main/java/com/longport/quote/OptionVolumeStats.java
@@ -1,6 +1,11 @@
 package com.longport.quote;
 
 public class OptionVolumeStats {
-    public String c;
-    public String p;
+    public String symbol;
+    public long callVolume;
+    public long putVolume;
+    public long callOpenInterest;
+    public long putOpenInterest;
+    public double pcVol;
+    public double pcOi;
 }